test: run LSP7 tests for proxy in separate test suite
