+++ package.json
@@ -49,7 +49,8 @@
     "test:lsp4": "hardhat test --no-compile tests/LSP4DigitalAssetMetadata/LSP4Compatibility.test.ts",
     "test:lsp6": "hardhat test --no-compile tests/LSP6KeyManager/LSP6KeyManager.test.ts",
     "test:lsp6init": "hardhat test --no-compile tests/LSP6KeyManager/LSP6KeyManagerInit.test.ts",
-    "test:lsp7": "hardhat test --no-compile tests/LSP7DigitalAsset/*.test.ts tests/LSP7DigitalAsset/extensions/*.test.ts",
+    "test:lsp7": "hardhat test --no-compile tests/LSP7DigitalAsset/standard/*.test.ts",
+    "test:lsp7init": "hardhat test --no-compile tests/LSP7DigitalAsset/proxy/*.test.ts",
     "test:lsp8": "hardhat test --no-compile tests/LSP8IdentifiableDigitalAsset/*.test.ts tests/LSP8IdentifiableDigitalAsset/extensions/*.test.ts",
     "test:lsp9": "hardhat test --no-compile tests/LSP9Vault/LSP9Vault.test.ts",
     "test:lsp9init": "hardhat test --no-compile tests/LSP9Vault/LSP9VaultInit.test.ts",
