+++ tests/LSP7DigitalAsset/LSP7CappedSupply.behaviour.ts
@@ -1,7 +1,7 @@
 import { SignerWithAddress } from "@nomiclabs/hardhat-ethers/signers";
 import { ethers } from "hardhat";
 import { expect } from "chai";
-import { LSP7CappedSupplyTester } from "../../../types";
+import { LSP7CappedSupplyTester } from "../../types";
 
 import type { BigNumber } from "ethers";
